// src/generated/schemas/-v2-site/-v2-site.zod.ts
export const getSiteResponse = zod.object({
  "id": zod.number().optional(),
  "uid": zod.string().optional().describe('Unique alphanumeric UID of this site'),
  "accountUid": zod.string().optional().describe('Unique alphanumeric UID of the account to which this site belongs'),
  "name": zod.string().optional(),
  "description": zod.string().optional(),
  "notes": zod.string().optional(),
  "onDemand": zod.boolean().optional(),
  "splashtopAutoInstall": zod.boolean().optional(),
  "proxySettings": zod.object({
  "host": zod.string().optional(),
  "password": zod.string().optional(),
  "port": zod.number().optional(),
  "type": zod.enum(['http', 'socks4', 'socks5']).optional(),
  "username": zod.string().optional()
}).optional().describe('Site\'s Proxy Settings'),
  "devicesStatus": zod.object({
  "numberOfDevices": zod.number().optional(),
  "numberOfOnlineDevices": zod.number().optional(),
  "numberOfOfflineDevices": zod.number().optional()
}).optional().describe('Devices Status'),
  "autotaskCompanyName": zod.string().optional(),
  "autotaskCompanyId": zod.string().optional(),
  "portalUrl": zod.string().optional()
}).describe('Site data')
// ...
export const updateProxyResponse = zod.object({
  "generalSettings": zod.object({
  "name": zod.string().optional(),
  "uid": zod.string().optional(),
  "description": zod.string().optional(),
  "onDemand": zod.boolean().optional()
}).optional().describe('General site settings data'),
  "proxySettings": zod.object({
  "host": zod.string().optional(),
  "password": zod.string().optional(),
  "port": zod.number().optional(),
  "type": zod.enum(['http', 'socks4', 'socks5']).optional(),
  "username": zod.string().optional()
}).optional().describe('Site\'s Proxy Settings'),
  "mailRecipients": zod.array(zod.object({
  "name": zod.string().optional(),
  "email": zod.string().optional(),
  "type": zod.string().optional(),
  "receives": zod.number().optional(),
  "receivesReports": zod.boolean().optional(),
  "receivesNewDevices": zod.boolean().optional(),
  "receivesAlerts": zod.boolean().optional()
}).describe('Mail recipient data')).optional()
}).describe('Site settings data')
// ...
export const getSiteSettingsResponse = zod.object({
  "generalSettings": zod.object({
  "name": zod.string().optional(),
  "uid": zod.string().optional(),
  "description": zod.string().optional(),
  "onDemand": zod.boolean().optional()
}).optional().describe('General site settings data'),
  "proxySettings": zod.object({
  "host": zod.string().optional(),
  "password": zod.string().optional(),
  "port": zod.number().optional(),
  "type": zod.enum(['http', 'socks4', 'socks5']).optional(),
  "username": zod.string().optional()
}).optional().describe('Site\'s Proxy Settings'),
  "mailRecipients": zod.array(zod.object({
  "name": zod.string().optional(),
  "email": zod.string().optional(),
  "type": zod.string().optional(),
  "receives": zod.number().optional(),
  "receivesReports": zod.boolean().optional(),
  "receivesNewDevices": zod.boolean().optional(),
  "receivesAlerts": zod.boolean().optional()
}).describe('Mail recipient data')).optional()
}).describe('Site settings data')