// src/generated/types/deviceNetworkInterface.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { DevicesType } from './devicesType';
import type { NetworkInterface } from './networkInterface';

/**
 * Device Network Interface data
 */
export interface DeviceNetworkInterface {
  id?: number;
  uid?: string;
  siteId?: number;
  siteUid?: string;
  siteName?: string;
  deviceType?: DevicesType;
  hostname?: string;
  intIpAddress?: string;
  extIpAddress?: string;
  nics?: NetworkInterface[];
}