// src/generated/types/filter.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { FilterType } from './filterType';

/**
 * Filter data
 */
export interface Filter {
  id?: number;
  name?: string;
  description?: string;
  type?: FilterType;
  dateCreate?: string;
  lastUpdated?: string;
}