// src/generated/types/getSitesParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */

export type GetSitesParams = {
page?: number;
max?: number;
/**
 * Optional. Filters results based on the provided value using the LIKE operator. Partial matches are allowed.
 */
siteName?: string;
};