// src/generated/types/site.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { ProxySettings } from './proxySettings';
import type { DevicesStatus } from './devicesStatus';

/**
 * Site data
 */
export interface Site {
  id?: number;
  /** Unique alphanumeric UID of this site */
  uid?: string;
  /** Unique alphanumeric UID of the account to which this site belongs */
  accountUid?: string;
  name?: string;
  description?: string;
  notes?: string;
  onDemand?: boolean;
  splashtopAutoInstall?: boolean;
  proxySettings?: ProxySettings;
  devicesStatus?: DevicesStatus;
  autotaskCompanyName?: string;
  autotaskCompanyId?: string;
  portalUrl?: string;
}